Duplication Code - Remove repetitive authorization header annotation
Code duplication in this case refers to the repetition of the @Headers("Authorization: token $key") annotation in each endpoint method of the "UserService" interface. This duplication can lead to maintenance issues and decreased code readability. The solution can be by adding the token in the interceptor in the configuration of retrofit. This eliminates the need for repeated annotations, resulting in cleaner and more maintainable code.