Indecent Exposure - Remove exposed token
The API key is hardcoded directly in the class. This is not recommended as it exposes sensitive information and makes it difficult to manage or update the key in the future. It would be better to use a more secure approach to handle the API key. One of the solution to hide the API key is using environment variables, in this case local.properties. Also make sure local.properties is included in .gitignore to prevent the it from uploaded to the github. This way, the API key will not be exposed and prevent another party to use the token exposed in the class.