Invalid names - Change adapter class name
The name "FollowanceAdapter" may not accurately reflect the purpose or functionality of the adapter, leading to confusion for other developers. To improve clarity and understanding, it is recommended to use more appropriate and meaningful names for the adapter class. Suggestions for alternative names could be "SocialAdapter", "FollowAdapter", or "FriendshipAdapter". Choosing a name that accurately describes the adapter's role will enhance code readability and maintainability. In this case, the adapter renamed to "UserFollowAdapter" as it represents "Following" and "Followers" of the user.