Invalid names - Change user following/followers layout xml name
The naming of the xml file "followance_row.xml" can also be considered as an invalid name. The name "followance_row.xml" is not intuitive and may cause confusion for other developers who are trying to understand the purpose or functionality of the layout file. The naming should be more appropriate and meaningful name for the xml. The recommended names would be "social_row.xml", "follow_row.xml", "friendship_row.xml". In this case with an addition of the word User, the name of the xml file would be "item_user_follow.xml" as it self-explains that this is the list layout of the user follow, could be "Following" and "Followers".