Speculative Generality - Remove unused code
The code is referred to as "Speculative Generality" because it includes an interface and a corresponding setter method setOnItemClickCallback that are not currently being used by the rest of the code. The OnItemClickCallback interface and the setOnItemClickCallback method provide a way to set a callback for item clicks in the ListUserAdapter. However, if this functionality is not currently used or planned to be used, it introduces unnecessary complexity and bloat to the code.